* branch-merge-for-wfr-ifs-3.12-0.9-283:
  IB/hfi1/tests: Add verbs directory to cron mail log
  IB/hfi1: Allow tuning of SDMA interrupt rate
  IB/hfi1: Fix checkpatch, klockwork, sparse issues
  staging: hfi1: Kconfig: remove 'CONFIG_' prefix

Conflicts:
	test/perf_log.cron
	test/tests/klocwork.whitelist
	test/tests/sparse.whitelist

+++ user_pages.c
@@ -50,8 +50,9 @@
 #include <linux/mm.h>
 #include <linux/sched.h>
 #include <linux/device.h>
+#include "hfi.h"
 
-int hfi1_acquire_user_pages(unsigned long vaddr, int npages, bool writable,
+int hfi1_acquire_user_pages(unsigned long vaddr, size_t npages, bool writable,
 			    struct page **pages)
 {
 	unsigned long pinned, lock_limit = rlimit(RLIMIT_MEMLOCK) >> PAGE_SHIFT;
